+++ users/RELEASES.md
@@ -49,3 +49,13 @@ Based on:
 - [typescript v2.0.3] users
 ### Releases
 - [NPM v2.0.3] https://www.npmjs.com/package/@wingspan/users/v/2.0.3 - users
+
+## 2023-12-11 01:31:05
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470eea6a7577e003be20a26
+- Speakeasy CLI 1.125.2 (2.210.6) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.0.4] users
+### Releases
+- [NPM v2.0.4] https://www.npmjs.com/package/@wingspan/users/v/2.0.4 - users